Data.Faceted

Description

Faceted Values with Faceted Execution.

Inspired by Thomas H. Austin and Cormac Flanagan, "Multiple Facets for Dynamic Information Flow."

Documentation

data Faceted ctx val

Faceted Value: < k ? V_H : V_L>

Principal is modeled by (ctx -> Bool). True means observer is private one, false means private one.

Note that faceted value can be nested: < k1 ? < k2 ? V_H : V_L> : < k3 ? W_H : W_L> >

Constructors

Faceted (ctx -> Bool) val val

principal :: Faceted ctx val -> ctx -> Bool

principal extractor

runFaceted :: Faceted ctx val -> ctx -> val

Evaluator for faceted values. Its semantics is faceted execution denoted in the paper above.

>>> runFaceted (Faceted (const True) True False) 1
True
>>> runFaceted (Faceted (const False) True False) 0
False

observe :: Faceted ctx val -> ctx -> val

an idiom of runFaceted

data FacetedT ctx m val

FacetedT: Faceted Value which can be a monad transformer.

Constructors

FacetedT (m (Faceted ctx val))

runFacetedT :: Functor m => FacetedT ctx m val -> ctx -> m val

FacetedT executor which also support faceted execution

observeT :: Functor m => FacetedT ctx m val -> ctx -> m val

an idiom for runFacetedT