Data.Faceted

Description

Faceted values with faceted execution.

Inspired by Thomas H. Austin and Cormac Flanagan, "Multiple Facets for Dynamic Information Flow."

Documentation

type Faceted ctx a = Free (Facets ctx) a

Faceted Value: < k ? V_H : V_L>

Principal is modeled by (ctx -> Bool). True means observer is private one, false means private one.

Facted values can be declared in very intuitive manner: (\x -> x > 0) ? 1 .: 0

Use (??) operator instead of (?) when you declare nested faceted values: (\x -> x > 0) ?? ((\y -> y > 1) ? 4 .: 3) .: ((\z -> z < 2) ? 2 .: 1)

(?) :: (ctx -> Bool) -> (a, a) -> Faceted ctx a

Constructor for unnested faceted values e.g. (x -> x > 0) ? 2 .: 1)

(??) :: (ctx -> Bool) -> (Faceted ctx a, Faceted ctx a) -> Faceted ctx a

Constructor for nested faceted values e.g. (\x -> x > 0) ?? ((\y -> y > 1) ? 4 .: 3) .: ((\z -> z < 2) ? 2 .: 1)

(.:) :: a -> a -> (a, a)

This works as a part of faceted values constructor.

principal :: Facets ctx val -> ctx -> Bool

principal extractor

runFaceted :: Faceted ctx a -> ctx -> a

observe :: Faceted ctx a -> ctx -> a

an idiom of runFaceted